split print and download protractor page objects
